refactor: split RPCClient to internal
